+++ api/src/main/java/marquez/db/models/ColumnLineageNodeData.java
@@ -19,7 +19,7 @@ public class ColumnLineageNodeData implements NodeData {
   @NonNull String dataset;
   @Nullable UUID datasetVersion;
   @NonNull String field;
-  @NonNull String fieldType;
+  @Nullable String fieldType;
   String transformationDescription;
   String transformationType;
   @NonNull List<InputFieldNodeData> inputFields;